Minor fixes to the setup process.
- Update scripts to use /bin/bash
- Remove exit on error and properly export DOTFILES_DIR

+++ install/vscode.sh
@@ -1,6 +1,11 @@
-#!/bin/sh
+#!/bin/bash
 
-curl -sL https://go.microsoft.com/fwlink/?LinkID=760868 -o /tmp/vscode.deb | sudo -E dpkg -i /tmp/vscode.deb
-sudo apt-get update
-sudo apt-get install -f
+curl -sL https://go.microsoft.com/fwlink/?LinkID=760868 -o /tmp/vscode.deb
+if [[ -f /tmp/vscode.deb ]]; then
+	sudo -E dpkg -i /tmp/vscode.deb
+	sudo apt-get update
+	sudo apt-get install -f
+else
+	echo "Could not download vscode"
+fi
